+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * improved `MeCms\Controller\Admin\LogController`, added `_read()` method and
     removed the `viewSerialized` action (the `view` action can use the 
     `view_as_serialized` template);
+* added `MeCms\Model\Table\UsersTable\beforeMarshal()` method;
 * added `MeCms\Controller\Traits\CheckLastSearchTrait` trait;
 * added `MeCms\Controller\Traits\DownloadTrait` trait;
 * added `MeCms\Model\Entity\Traits\PreviewAccessorTrait` trait;